requirements.txt
anyio==3.6.2
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.94.0
fastapi-jwt-auth==0.5.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
huesdk==1.4
idna==3.4
ifaddr==0.2.0
imutils==0.5.4
numpy==1.24.2
opencv-python==4.7.0.72
psycopg2-binary==2.9.6
pydantic==1.10.6
PyJWT==1.7.1
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.26.0.post1
stripe==5.2.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
zeroconf==0.47.3
